fix: mudança de pasta e refatora a parte de busca
