--- src/queries/busca.js
@@ -1,7 +1,0 @@
-class BuscaStrategy {
-    buscar(Contatos, termo) {
-        throw new Error("Este método deve ser implementado.");
-    }
-}
-
-module.exports = BuscaStrategy;
