+++ src/services/BuscaStrategy.js
@@ -1,0 +1,9 @@
+class BuscaStrategy {
+  buscar(contatos, termo) {
+    throw new Error(
+      "O método 'buscar' deve ser implementado na subclasse de BuscaStrategy."
+    );
+  }
+}
+
+module.exports = BuscaStrategy;
